@@ -8,23 +8,20 @@
     "@fortawesome/free-regular-svg-icons": "^6.5.2",
     "@fortawesome/free-solid-svg-icons": "^6.5.2",
     "@fortawesome/react-fontawesome": "^0.2.0",
+    "@reduxjs/toolkit": "^2.2.4",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
-    "@types/jest": "^26.0.15",
-    "@types/node": "^12.0.0",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/react-helmet": "^6.1.4",
-    "@types/react-router-dom": "^5.3.1",
-    "@types/styled-components": "^5.1.15",
+    "antd": "^5.17.2",
     "caniuse-lite": "^1.0.30001270",
     "formik": "^2.2.6",
     "gh-pages": "^3.1.0",
-    "react": "^17.0.1",
+    "js-cookie": "^3.0.5",
+    "react": "^18.3.1",
     "react-awesome-reveal": "^3.7.0",
-    "react-dom": "^17.0.1",
+    "react-dom": "^18.3.1",
     "react-helmet": "^6.1.0",
+    "react-redux": "^9.1.2",
     "react-router-dom": "^5.3.0",
     "react-scripts": "4.0.2",
     "sass": "^1.43.2",
@@ -33,13 +30,35 @@
     "typescript": "^4.1.2",
     "web-vitals": "^1.0.1"
   },
+  "devDependencies": {
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "@openapitools/openapi-generator-cli": "^2.13.4",
+    "@types/jest": "^26.0.15",
+    "@types/js-cookie": "^3.0.6",
+    "@types/node": "^12.0.0",
+    "@types/react": "^18.3.2",
+    "@types/react-dom": "^18.3.0",
+    "@types/react-helmet": "^6.1.4",
+    "@types/react-router-dom": "^5.3.1",
+    "@types/styled-components": "^5.1.15",
+    "dotenv": "^16.4.5",
+    "dotenv-cli": "^7.4.2",
+    "prettier": "^3.2.5"
+  },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
     "predeploy": "npm run build",
-    "deploy": "gh-pages -d build"
+    "deploy": "gh-pages -d build",
+    "lint": "eslint . --ext .js,.jsx,.ts,.tsx",
+    "lint:fix": "eslint --fix ./src --ext .jsx,.ts,.tsx",
+    "prettier": " prettier \"src/**/*.{ts,tsx}\"",
+    "prettier:write": " prettier --write \"src/**/*.{ts,tsx}\"",
+    "openapi:download": "curl \"http://localhost:4000/api-json\" -o ./src/shared/types/api/openapi.json",
+    "openapi:generate": "openapi-generator-cli generate -i ./src/shared/types/api/openapi.json -g typescript-axios --skip-validate-spec --enable-post-process-file -o ./src/shared/types/api/generated",
+    "openapi": "yarn openapi:download && yarn openapi:generate"
   },
   "eslintConfig": {
     "extends": [
